repo: introduce apt-stage-cache to prevent unnecessary apt fetches
It several seconds for the apt cache to to be updated and several
more to re-verify the hashes of archives that were already fetched.
It may even take minutes depending on performance and how many
stage-packages are being requested.

This introduces AptStageCaheDb to save off successful calls
to install_stage_packages() and its associated results, in a
manner persistent across runs.

Now install_stage_packages() will check with the db to see if
we have any cached requests that were previously successful, and
re-use those results.

If the db is incompatible or invalid in any way, we simply clear
the cache, resulting in new fetches.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>